Move reference computation to utils class
